+++ R/render_site.R
@@ -24,8 +24,6 @@
 #' @importFrom cli cli_h1 cli_h2 cli_abort cli_text cli_alert_warning
 #' @importFrom tictoc tic toc
 #' @importFrom servr daemon_stop httw
-#' @importFrom future plan
-#' @importFrom future.mirai mirai_multisession
 #' @importFrom quarto quarto_render
 #' @importFrom gert git_status git_remote_list
 #' @export
@@ -58,7 +56,6 @@ render_site <- function(
 
   tictoc::tic()
   servr::daemon_stop()
-  future::plan(future.mirai::mirai_multisession, workers = workers)
 
   if(fs::dir_exists("_site"))
     tryCatch(