+++ R/render_wp.R
@@ -29,8 +29,6 @@
 #' @importFrom cli cli_h1 cli_h2 cli_abort cli_text cli_alert_warning cli_alert_success
 #' @importFrom tictoc tic toc
 #' @importFrom servr daemon_stop httw
-#' @importFrom future plan
-#' @importFrom future.mirai mirai_multisession
 #' @importFrom quarto quarto_render
 #' @export
 render_wp <- function(
@@ -78,7 +76,6 @@ render_wp <- function(
 
   tictoc::tic()
   servr::daemon_stop()
-  future::plan(future.mirai::mirai_multisession, workers = workers)
 
   # ---- 2.5. registre central (détermine stage avant le rendu) --------------
   # Consulte ofceweb/wp-registry (disposition wp/index.json + wp/{année}.json)